+++ Source/UnitTests/Common/CMakeLists.txt
@@ -7,6 +7,7 @@ add_dolphin_test(CommonFuncsTest CommonFuncsTest.cpp)
 add_dolphin_test(EventTest EventTest.cpp)
 add_dolphin_test(FixedSizeQueueTest FixedSizeQueueTest.cpp)
 add_dolphin_test(FlagTest FlagTest.cpp)
+add_dolphin_test(FloatUtilsTest FloatUtilsTest.cpp)
 add_dolphin_test(MathUtilTest MathUtilTest.cpp)
 add_dolphin_test(NandPathsTest NandPathsTest.cpp)
 add_dolphin_test(SPSCQueueTest SPSCQueueTest.cpp)
